// src/app/core/models/client.model.ts
export class Client {
    userNumber:string;
    inmateNumber: string;
    housingUnit: string;
    callingProfile: string;
    userGroup: string;
    
    lastName: string;
    middletName: string;
    firstName: string;
    country: string;
    
    timeZone?: string;
    status: string;
    lock: boolean;
    lockPin: string;
    credit?: number;
}